requirements.txt
functions-framework==3.*
ollama
requests
supabase
python-dotenv
google-cloud-vision
google-genai